Add support for pods and MP3s
Not all programs are considered broadcasts, instead they are available
as pods to download for a larger amount of time than usual.

Add support for such pods. As these pods typically are made available
for download as an MP3, add MP3 support - assuming the files already
contain the right metadata (as seems to be the case).